basic routing with Example
