+++ src/AboutOne.js
@@ -1,0 +1,10 @@
+const AboutOne =()=>{
+    return(
+        <div>
+          <h1>About Page</h1>
+          <p>This is About Page</p>
+        </div>
+      );
+}
+
+export default AboutOne;
