+++ src/HomeOne.js
@@ -1,0 +1,10 @@
+const HomeOne=()=>{
+    return(
+        <div>
+          <h1>Home Page</h1>
+          <p>This is Home Page</p>
+        </div>
+    );
+}
+
+export default HomeOne;
