+++ src/Nav.js
@@ -1,0 +1,13 @@
+import { Link } from "react-router-dom";
+
+const Nav=()=>{
+    return(
+        <div>
+            <Link to='/'>Home</Link>
+            <Link to='/about'>About</Link>
+            {/*<a href="/about">About</a>*/}
+        </div>
+    );
+}
+
+export default Nav;
